Split image initial job transaction
